Refactored cython code.
