--- pylcogt/utils/stats.pyx
@@ -1,215 +1,0 @@
-# cython: boundscheck=False, nonecheck=False, wraparound=False
-# cython: cdivision=True
-from __future__ import absolute_import, print_function, division
-import numpy as np
-from libc.stdint cimport uint8_t
-from libc.stdlib cimport malloc
-cimport numpy as np
-
-cimport cython
-from cython.parallel import parallel, prange
-
-np.import_array()
-
-
-__author__ = 'cmccully'
-
-cdef extern from "median_utils.h":
-    float quick_select(float * k, int k, int n) nogil
-    void median2d(float * data, float * output, uint8_t * mask, int nx, int ny) nogil
-
-
-@cython.boundscheck(False)
-@cython.wraparound(False)
-def _quick_select(float[::1] a not None, int k):
-    cdef float value
-    cdef int size = a.size
-    with nogil:
-        value = quick_select(&a[0], k, size)
-    return value
-
-
-def median(d, axis=None, mask=None):
-    """median(d, axis=None, mask=None)\n
-    Find the median of a numpy array. If an axis is provided, then find the median along
-    the given axis. If mask is included, elements in d that have a non-zero mask value will
-    be ignored.
-
-    Parameters
-    ----------
-    d : float32 numpy array
-        Input array to find the median.
-
-    axis : int
-           Index of the array to take the median
-    mask: unit8 numpy array
-          Numpy array of bitmask values. Non-zero values are ignored when calculating the median.
-
-    Returns
-    -------
-    med : float32 numpy array
-        The median value. If axis is None, then we return a single float.
-
-    Notes
-    -----
-    Makes extensive use of the quick select algorithm written in C, included in median_utils.c.
-    If all of the elements in the array are masked (or all of the elements of the axis of interest
-    are masked), we return zero.
-    """
-    if axis is None:
-        if mask is not None:
-            median_mask = mask.ravel()
-        else:
-            median_mask = np.zeros(d.size, dtype=np.uint8)
-        med = _median1d(d.ravel().astype('f4'), median_mask.astype(np.uint8))
-    else:
-        output_shape = np.delete(d.shape, axis)
-        nx = d.shape[axis]
-        ny = d.size // nx
-
-        if mask is not None:
-            median_mask = np.swapaxes(mask, axis, -1).reshape(ny, nx).astype(np.uint8)
-        else:
-            median_mask = np.zeros((ny, nx), dtype=np.uint8)
-
-        med = _median2d(np.swapaxes(d, axis, -1).reshape(ny, nx).astype('f4'),
-                        mask=median_mask)
-        med = np.asarray(med).reshape(output_shape)
-
-    return med
-
-
-@cython.boundscheck(False)
-@cython.wraparound(False)
-cdef float _cmedian1d(float* ptr, int n) nogil:
-    cdef float med = 0.0
-    cdef int k = (n - 1) // 2
-    if n > 0:
-        med = quick_select(ptr, k, n)
-        if n % 2 == 0:
-            med += quick_select(ptr, k + 1, n)
-            med /= 2.0
-    return med
-
-
-@cython.boundscheck(False)
-@cython.wraparound(False)
-def _median1d(float[::1] d not None, uint8_t[::1] mask not None):
-    """_median1d(d, mask)\n
-    Find the median of a numpy array. If an axis is provided, then find the median along
-    the given axis. If mask is included, elements in d that have a non-zero mask value will
-    be ignored.
-
-    Parameters
-    ----------
-    d : float numpy array
-        Input array to find the median.
-
-    mask: unit8 numpy array
-          Numpy array of bitmask values. Non-zero values are ignored when calculating the median.
-
-    Returns
-    -------
-    med : float
-        The median value.
-
-    Notes
-    -----
-    Makes extensive use of the quick select algorithm written in C, included in median_utils.c.
-    If all of the elements in the array are masked (or all of the elements of the axis of interest
-    are masked), we return zero. This has comparable performance to np.median on unmasked data,
-    but does not require the gil. For masked arrays, the performance is significantly better
-    (anecdotally, I have seen improvements of more than order of magnitude, but I have not done
-    a comprehensive benchmark).
-    """
-
-    cdef int n = d.shape[0]
-
-    cdef float[::1] median_array = np.empty(n, dtype=np.float32)
-
-    cdef int n_unmasked_pixels = 0
-    cdef int i = 0
-    for i in range(n):
-        if mask[i] == 0:
-            median_array[n_unmasked_pixels] = d[i]
-            n_unmasked_pixels += 1
-
-    return _cmedian1d(&median_array[0], n_unmasked_pixels)
-
-
-@cython.boundscheck(False)
-@cython.wraparound(False)
-def _median2d(float[:, ::1] d, uint8_t[:, ::1] mask):
-
-    cdef int nx = d.shape[1]
-    cdef int ny = d.shape[0]
-
-    cdef int j = 0
-    cdef int i
-
-    cdef float[::1] output_array = np.empty(ny, dtype=np.float32)
-    cdef float* median_array
-    cdef int n_unmasked_pixels = 0
-    median_array =<float *> malloc(nx * sizeof(float))
-    with nogil, parallel():
-        for j in prange(ny):
-            n_unmasked_pixels = 0
-            for i in range(nx):
-                if mask[j, i] == 0:
-                    median_array[n_unmasked_pixels] = d[j, i]
-                    n_unmasked_pixels = n_unmasked_pixels + 1
-
-            output_array[j] = _cmedian1d(median_array, n_unmasked_pixels)
-
-    return output_array
-
-
-def absolute_deviation(a, axis=None, mask=None):
-    # Calculate the image deviation
-    a_median = median(a, axis=axis, mask=mask)
-    if axis is not None:
-        a_median = np.expand_dims(a_median, axis=axis)
-    return np.abs(a - a_median)
-
-
-def median_absolute_deviation(a, axis=None, abs_deviation=None, mask=None):
-    if abs_deviation is None:
-        abs_deviation = absolute_deviation(a, axis=axis, mask=mask)
-
-    return median(abs_deviation, axis=axis, mask=mask)
-
-
-def robust_standard_deviation(a, axis=None, abs_deviation=None, mask=None):
-    return 1.4826 * median_absolute_deviation(a, axis=axis, abs_deviation=abs_deviation, mask=mask)
-
-
-def sigma_clipped_mean(a, sigma, axis=None, mask=None):
-    """
-    """
-    abs_deviation = absolute_deviation(a, axis=axis, mask=mask)
-
-    robust_std = robust_standard_deviation(a, axis=axis, abs_deviation=abs_deviation, mask=mask)
-
-    if axis is not None:
-        robust_std = np.expand_dims(robust_std, axis=axis)
-
-    # Throw away any values that are N sigma from the median
-    sigma_mask = (abs_deviation > (sigma * robust_std))
-
-    if mask is not None:
-        sigma_mask = np.logical_or(sigma_mask, (mask > 0))
-
-    data_copy = a.copy()
-    data_copy[sigma_mask] = 0.0
-
-    # Take the sigma clipped mean
-    mean_values = data_copy.sum(axis=axis)
-
-    n_good_pixels = (~sigma_mask).sum(axis=axis)
-    if axis is None:
-        if n_good_pixels > 0:
-            mean_values /= n_good_pixels
-    else:
-        mean_values[n_good_pixels > 0] /= n_good_pixels[n_good_pixels > 0]
-
-    return mean_values
